

\name{ggplot_geopos}
\alias{ggplot_geopos}
\alias{plot_geopos}

\alias{get_geopos}

\title{
reads and plots geolocation estimates derived from archival tagging data
}
\description{
In case that geolocations are provided by csv-files or data frames, line and scatter plots are implemented.
If ncdf-files or kmz-files, generated by the \href{http://my.wildlifecomputers.com/}{Wildlife Computers-data portal}, are selected, a \link{SpatialPolygonsDataFrame} will be created and a surface probability maps are illustrated. The netcdf transformation procedure is based on the R-code given in the \href{https://static.wildlifecomputers.com/manuals/Location-Processing-User-Guide.pdf}{location processing user guide} by \href{http://wildlifecomputers.com/}{Wildlife Computers}. The kmz-files already include the contour lines of the 50, 95 and 99\% likelihood areas that are being extracted and likewise transformed to a \link{SpatialPolygonsDataFrame}. In case of kmz-files, no other areas can be selected.
}
\usage{

ggplot_geopos(x, ggobj, xlim, ylim, zlim, standard_year=FALSE,
              date_format, lang_format="en", tz="UTC", cb.date_format, 
              cbpos, cb.height = 10, cb.xlab = "", prob_lim=.75, 
              pal="jet", alpha=70, type="p", 
              main="", lwd=1, size=2, shape=19, ...)
            
get_geopos(x, xlim, ylim, date_format, lang_format="en", tz="UTC", 
           proj4string, prob_lim=.5)
}



\arguments{
\item{x}{

\link{data.frame} containing horziontal position records (allowed column names are 'Most.Likely.Longitude', 'Longitude' or 'Lon' and 'Most.Likely.Latitude', 'Latitude' or 'Lat', respectively. path and file name of \code{.csv}, \code{.kmz}, \code{.kml} or \code{.nc}-files, or a loaded nc-file via \code{get_geopos}.

Please note that netcdf- and kmz-file outputs from similiar probability thresholds (e.g. 0.50) may differ due to differences in the generation algorithm.

Please also note that the kmz-files from WC include only a subsample (up to 50) of the GPE3 likelihood areas. The kmz-file transformation is very fast, but due to the limitation stated above, only recommended for display purposes. Please contact WC if you need the complete GPE3 likelihood areas or use the netcdf-file transformation.
}

\item{ggobj}{
ggplot object.
}

%\item{v_area}{
%Character string used as a short key by the \link{oceanmap}-package to identify region (limits) to be plotted. See   \link{region_definitions} for existing for region definitions and use add.region to define new ones.
%}

\item{xlim, ylim}{
Numeric vector, defining the limts of the x and y-axes.
}

\item{zlim, standard_year}{
date range of the colorbar. If standard_year is set \code{TRUE}, positions are standardized on a yearly basis.
}

\item{date_format, lang_format, tz}{
character strings indicating the date format, language format and the corresponding time zone, defined by the vectors Date and Time (by default: date_format="\%d-\%b-\%Y \%H:\%M:\%S", lang_format="en", tz='UTC')
If formatting fails, please check as well the input language format, defined by \code{lang_format} (and use abbrviations such as "en" for English,"es" for Spanish, "fr" for French, etc.) as well.
}

\item{proj4string}{
Coordinate reference system (CRS; \link{projection}).
}

\item{cb.date_format}{
character strings indicating the date format of the color bar ticks (by default "\%Y-\%m-\%d").
}

\item{cbpos, cb.xlab, cb.height}{
position, title and height of the colorbar
}

\item{prob_lim}{
in case that a kmz, kml, or netcdf-file (.nc) is selected, the value defines the limit of the probability surfaces in \% (By default 0.50 for 50\%). Note that in case of kmz, kml-files valid values are 0.50, 0.95 or 0.99). Otherwise ignored.
}

\item{pal}{
  color map to be plotted in case of polygon (.nc-files) or scatter plots (default is the 'jet'-colormap). See \link{cmap} for pre-installed color maps. Note that tracking data with constant time steps is being assumed in the color assignment. To verify this, a \link{data.frame} containing the colors at each time steps will be returned for polygon and scatter plots.
}


\item{alpha}{transparency of polygons and dots to be plotted in percent (By default 70\%).
}

\item{type}{
character string giving the type of plot desired. The following values are possible, for details (By default "p" for points, but "l" for lines is also implemented).
}

\item{size, shape}{
size and dot-type (by default '19' for solid dots) of the points to be plotted (requires \code{'type'} set to points or line-points).
}

\item{lwd}{
line width
}

\item{\dots}{
additional arguments to be passed to \link{plotmap}.
}

\item{main}{
an overall title for the plot
}

}



\seealso{\link{plotmap}, \link{plot_DepthTS}, \link{hist_tat}, \link{hist_tad}} 


\author{
Robert K. Bauer
}

\examples{
# ## example 1a) line plot from several csv-files:
# csv_file <- system.file("example_files/15P1019-104659-1-GPE3.csv",package="RchivalTag")
# pos <- get_geopos(csv_file)  ## show tracks as line plot
# ggplot_geopos(pos)
# ggplot_geopos(pos,type = "l")
# 
# ggplot_geopos(pos,zlim=c("2016-08-01","2016-09-01"))  ## change dates of colorbar
# ggplot_geopos(pos,standard_year = T,zlim=c("2016-08-01","2016-09-01"))  ## standardize years
# 
# ## load second file and add to plot:
# csv_file2 <- system.file("example_files/14P0911-46177-1-GPE3.csv",package="RchivalTag")
# pos2 <- get_geopos(csv_file2)  ## show tracks as line plot
# ggplot_geopos(pos2)
# 
# pos3 <- rbind(pos,pos2)
# ggplot_geopos(pos3,type = "l")
# 
# 
# ## example 1b) scatter plot from csv-file on existing landmask:
# ggobj <- ggplotmap('lion',grid.res = 5) # use keyword to derive area limits
# ggplot_geopos(csv_file,ggobj)
# 
# ## alternatives:
# pos <- get_geopos(csv_file)
# ggplot_geopos(pos,ggobj)
# 
# r <- oceanmap::regions("lion")
# ggplot_geopos(pos, xlim = r$xlim, ylim = r$ylim)
# 
# 
# ## example 2) probability surfaces of horizontal tracks from nc-file:
# ## this can take some time as it inlcudes time consuming data processing
# nc_file <- system.file("example_files/15P1019-104659-1-GPE3.nc",package="RchivalTag") 
# ggplot_geopos(nc_file)
# 
# ## alternative:
# pols_df <- get_geopos(nc_file)
# ggplot_geopos(pols_df)
# 
# 
# ## example 3) probability surfaces of horizontal tracks from kmz-file:
# kmz_file <- system.file("example_files/15P1019-104659-1-GPE3.kmz",package="RchivalTag")
# ggplot_geopos(kmz_file)
# 
# kmz_file2 <- system.file("example_files/15P0986-15P0986-2-GPE3.kmz",package="RchivalTag")
# ggplot_geopos(kmz_file2)
# 
# ## combine polygon tracks:
# k1 = get_geopos(kmz_file)
# k2 = get_geopos(kmz_file2)
# 
# ggobj <- ggplotmap("mednw4")
# p1 <- ggplot_geopos(k1,ggobj = ggobj,zlim = as.Date(range(c(k1$datetime,k2$datetime))))
# p2 <- ggplot_geopos(k2,p1,zlim = as.Date(range(c(k1$datetime,k2$datetime))))
# p2

}

